# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserLoginResponse": {
				"properties": {
					"aToken": {
						"type": "string"
					},
					"rToken": {
						"type": "string"
					}
				},
				"required": [
					"aToken",
					"rToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_UserLoginResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserLoginResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"properties": {
							"data": {
								"nullable": true
							},
							"reason": {
								"type": "string",
								"nullable": true
							},
							"errorCode": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_string_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"type": "string"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThreadType": {
				"enum": [
					"아티클",
					"팀원모집",
					"질문"
				],
				"type": "string"
			},
			"coffectChatCardDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"grade": {
						"type": "string"
					},
					"introduce": {
						"type": "string"
					},
					"categoryMatch": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"profileImage": {
						"type": "string"
					},
					"mail": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"grade",
					"introduce",
					"categoryMatch",
					"profileImage",
					"mail"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_coffectChatCardDTO_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/coffectChatCardDTO"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "coffect-be",
		"version": "1.0.0",
		"description": "Coffect BackEnd Repository",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "로그인성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_UserLoginResponse_"
								}
							}
						}
					},
					"400": {
						"description": "비밀번호 불일치",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC405",
												"reason": "비밀번호가 일치하지 않습니다",
												"data": "비밀번호가 일치하지 않습니다"
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 아이디",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC404",
												"reason": "존재하지 않는 아이디입니다",
												"data": "존재하지 않는 아이디입니다"
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "로그인",
				"summary": "로그인",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "유저아이디, 비밀번호",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"userId": {
										"type": "string"
									},
									"userPassword": {
										"type": "string"
									}
								},
								"required": [
									"userId",
									"userPassword"
								],
								"type": "object",
								"description": "유저아이디, 비밀번호"
							}
						}
					}
				}
			}
		},
		"/user/refresh": {
			"get": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "토큰 재발급",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_UserLoginResponse_"
								}
							}
						}
					},
					"400": {
						"description": "헤더 누락값 존재",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC404",
												"reason": "누락값이 존재합니다.",
												"data": "헤더에 토큰이 존재하지 않습니다."
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "jwt 토큰 만료",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "ERR-1",
												"reason": "Expired",
												"data": {
													"name": "TokenExpiredError",
													"message": "jwt expired",
													"expiredAt": "2025-07-14T01:29:15.000Z"
												}
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "일치하지 않는 토큰 사용",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "ERR-1",
												"reason": "JsonWebToken error",
												"data": "DB에 사용자 로그인 정보가 존재하지 않습니다. 다시 로그인해주세요"
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "ERR-1",
												"reason": "JsonWebToken error",
												"data": {
													"name": "JsonWebTokenError",
													"message": "invalid signature"
												}
											},
											"success": null
										}
									},
									"Example 3": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "ERR-1",
												"reason": "JsonWebToken error",
												"data": {
													"statusCode": 404,
													"code": "ERR-1",
													"description": "유효하지 않은 토큰입니다."
												}
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "데이터베이스에 토큰이 존재하는지 검증하고, 유효할 경우 새로운 토큰을 발급해준다.",
				"summary": "토큰검증",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"409": {
						"description": "중복된 아이디",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC409",
												"reason": "이미 중복된 아이디입니다.",
												"data": "아이디가 중복됨"
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "데이터베이스 삽입 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC500",
												"reason": "서버 오류가 발생했습니다.",
												"data": "데이터베이스 삽입에 실패했습니다."
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "ERR2",
												"reason": "서버에 사진을 올리는중 오류가 발생했습니다.",
												"data": "사진을 올리는 중 오류가 발생했습니다"
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "회원가입 form-data로 img, userInfo두가지를 따로 보낸다. 이때 userInfo는 json을 직렬화해 보내야하며 img에는 파일을 그대로 첨부해 보낸다.",
				"summary": "회원가입",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"profile": {
										"type": "string"
									},
									"userInfo": {
										"properties": {
											"interest": {
												"items": {
													"type": "number",
													"format": "double"
												},
												"type": "array"
											},
											"name": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"studentId": {
												"type": "number",
												"format": "double"
											},
											"major": {
												"type": "string"
											},
											"univ": {
												"type": "string"
											},
											"id": {
												"type": "string"
											},
											"password": {
												"type": "string"
											}
										},
										"required": [
											"interest",
											"name",
											"email",
											"studentId",
											"major",
											"univ",
											"id",
											"password"
										],
										"type": "object"
									}
								},
								"required": [
									"profile",
									"userInfo"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/idcheck": {
			"post": {
				"operationId": "Idcheck",
				"responses": {
					"200": {
						"description": "존재하지 않는 아이디",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_string_"
								}
							}
						}
					},
					"409": {
						"description": "중복된 아이디",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC409",
												"reason": "이미 중복된 아이디입니다.",
												"data": "아이디가 중복됨"
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "데이터베이스를 조회해 해당 아이디가 겹치는지 아닌지 확인한다.",
				"summary": "아이디 중복 체크",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/univ/cert": {
			"post": {
				"operationId": "Cert",
				"responses": {
					"200": {
						"description": "인증에 성공했습니다",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_string_"
								}
							}
						}
					},
					"401": {
						"description": "인증코드가 만료되었습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC1",
												"reason": "인증코드가 일치하지 않습니다.",
												"data": "인증코드가 일치하지 않습니다."
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC2",
												"reason": "인증코드가 유효하지 않습니다.",
												"data": "인증코드가 유효하지 않습니다."
											},
											"success": null
										}
									},
									"Example 3": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "EC3",
												"reason": "인증코드가 만료되었습니다,",
												"data": "인증코드가 만료되었습니다."
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "메일 인증번호가 일치하는지 확인한다",
				"summary": "이메일 인증번호 확인",
				"tags": [
					"Univ Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									},
									"certCode": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"email",
									"certCode"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/thread/add": {
			"post": {
				"operationId": "AddThread",
				"responses": {
					"201": {
						"description": "게시글 업로드 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_string_"
								}
							}
						}
					},
					"400": {
						"description": "유저 인증 정보가 없습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "ERR-1",
												"reason": "유저 인증 정보가 없습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "게시글 트랜잭션 처리에 실패했습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "THR-02",
												"reason": "게시글 업로드에 실패했습니다.",
												"data": null
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "THR-03",
												"reason": "게시글 트랜잭션 처리에 실패했습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "\n게시글 업로드 API",
				"summary": "게시글 업로드",
				"tags": [
					"Thread Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- 게시글 정보",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"threadSubject": {
										"type": "number",
										"format": "double"
									},
									"threadBody": {
										"type": "string"
									},
									"threadTitle": {
										"type": "string"
									},
									"type": {
										"$ref": "#/components/schemas/ThreadType"
									},
									"userId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"threadSubject",
									"threadBody",
									"threadTitle",
									"type"
								],
								"type": "object",
								"description": "- 게시글 정보"
							}
						}
					}
				}
			}
		},
		"/home/postTodayInterest": {
			"post": {
				"operationId": "PostTodayInterestController",
				"responses": {
					"200": {
						"description": "성공적으로 Data를 넣었습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_string_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HE400",
												"reason": "주제 선정해주세요.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HE500",
												"reason": "서버 오류가 발생했습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "Coffect coffeeChat Home API.",
				"summary": "매일 커피챗 추천 항목을 받는 API (가까운 거리 순 <1>, 나와 비슷한 관심사 <2>, 같은 학번 <3>, 최근에 글을 쓴 사람 <4>)",
				"tags": [
					"Home Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "유저 Token & 하루 관심사 정보 수정",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"todayInterest": {
										"type": "number",
										"format": "double"
									},
									"userId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"todayInterest",
									"userId"
								],
								"type": "object",
								"description": "유저 Token & 하루 관심사 정보 수정"
							}
						}
					}
				}
			}
		},
		"/home/getCardClose": {
			"get": {
				"operationId": "CardCloseCoffeeChatController",
				"responses": {
					"200": {
						"description": "성공적으로 Data를 넣었습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_coffectChatCardDTO_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HE400",
												"reason": "주제 선정해주세요.",
												"data": null
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HE401",
												"reason": "오늘 하루 추천 커피챗 횟수를 초과 했습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HE500",
												"reason": "서버 오류가 발생했습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "Coffect coffeeChat Home API.",
				"summary": "카드를 가져오는 API (가까운 거리 순 <1>, 나와 비슷한 관심사 <2>, 같은 학번 <3>, 최근에 글을 쓴 사람 <4>)",
				"tags": [
					"Home Controller"
				],
				"security": [],
				"parameters": []
			}
		},
		"/home/postSuggestCoffeeChat": {
			"post": {
				"operationId": "PostSuggestCoffeeChat",
				"responses": {
					"200": {
						"description": "성공적으로 Data를 넣었습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_string_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HE402",
												"reason": "내용이 누락되어있습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HE500",
												"reason": "서버 오류가 발생했습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "Coffect coffeeChat Home API.",
				"summary": "커피챗 제안 요청하는 API",
				"tags": [
					"Home Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"suggestion": {
										"type": "string"
									},
									"otherUserid": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"suggestion",
									"otherUserid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}